move consent to scopes to auth slice
