+++ src/app/services/slices/auth.slice.ts
@@ -1,10 +1,17 @@
-import { createSlice, PayloadAction } from '@reduxjs/toolkit';
+import { BrowserAuthError } from '@azure/msal-browser';
+import { MessageBarType } from '@fluentui/react';
+import { createAsyncThunk, createSlice, PayloadAction } from '@reduxjs/toolkit';
 
 import { authenticationWrapper } from '../../../modules/authentication';
-import { Mode } from '../../../types/enums';
+import { getConsentAuthErrorHint } from '../../../modules/authentication/authentication-error-hints';
 import { AppDispatch } from '../../../store';
-import { ApplicationState } from '../../../types/root';
 import { AuthenticateResult } from '../../../types/authentication';
+import { Mode } from '../../../types/enums';
+import { ApplicationState } from '../../../types/root';
+import { translateMessage } from '../../utils/translate-messages';
+import { fetchAllPrincipalGrants } from '../actions/permissions-action-creator';
+import { getProfileInfo } from './profile.slice';
+import { setQueryResponseStatus } from './query-status.slice';
 
 const initialState: AuthenticateResult = {
   authToken: {
@@ -53,6 +60,59 @@ export function signOut() {
     dispatch(signOutSuccess());
   };
 }
+
+const validateConsentedScopes = (scopeToBeConsented: string[], consentedScopes: string[],
+  consentedResponse: string[]) => {
+  if (!consentedScopes || !consentedResponse || !scopeToBeConsented) {
+    return consentedResponse;
+  }
+  const expectedScopes = [...consentedScopes, ...scopeToBeConsented];
+  if (expectedScopes.length === consentedResponse.length) {
+    return consentedResponse;
+  }
+  return expectedScopes;
+}
+
+export const consentToScopes = createAsyncThunk(
+  'auth/consentToScopes',
+  async (scopes: string[], { dispatch, getState }) => {
+    try {
+      const { profile, auth: { consentedScopes } } = getState() as ApplicationState;
+      const authResponse = await authenticationWrapper.consentToScopes(scopes);
+      if (authResponse && authResponse.accessToken) {
+        dispatch(getAuthTokenSuccess());
+        const validatedScopes = validateConsentedScopes(scopes, consentedScopes, authResponse.scopes);
+        dispatch(getConsentedScopesSuccess(validatedScopes));
+        if (
+          authResponse.account &&
+          authResponse.account.localAccountId !== profile?.user?.id
+        ) {
+          dispatch(getProfileInfo());
+        }
+        dispatch(
+          setQueryResponseStatus({
+            statusText: translateMessage('Success'),
+            status: translateMessage('Scope consent successful'),
+            ok: true,
+            messageType: MessageBarType.success
+          })
+        );
+        dispatch(fetchAllPrincipalGrants());
+      }
+    } catch (error: unknown) {
+      const { errorCode } = error as BrowserAuthError;
+      dispatch(
+        setQueryResponseStatus({
+          statusText: translateMessage('Scope consent failed'),
+          status: errorCode,
+          ok: false,
+          messageType: MessageBarType.error,
+          hint: getConsentAuthErrorHint(errorCode)
+        })
+      );
+    }
+  }
+);
 
 export function signIn() {
   return (dispatch: AppDispatch) => dispatch(getAuthTokenSuccess());
